+++ sdk/java/src/main/java/com/pulumi/dockerbuild/inputs/CacheToGitHubActionsArgs.java
@@ -277,7 +277,7 @@ public final class CacheToGitHubActionsArgs extends com.pulumi.resources.Resourc
         public CacheToGitHubActionsArgs build() {
             $.ignoreError = Codegen.booleanProp("ignoreError").output().arg($.ignoreError).def(false).getNullable();
             $.mode = Codegen.objectProp("mode", CacheMode.class).output().arg($.mode).def(CacheMode.Min).getNullable();
-            $.scope = Codegen.stringProp("scope").output().arg($.scope).env("buildkit").def("").getNullable();
+            $.scope = Codegen.stringProp("scope").output().arg($.scope).def("buildkit").getNullable();
             $.token = Codegen.stringProp("token").secret().arg($.token).env("ACTIONS_RUNTIME_TOKEN").def("").getNullable();
             $.url = Codegen.stringProp("url").output().arg($.url).env("ACTIONS_CACHE_URL").def("").getNullable();
             return $;